+++ software/ADPL_electron/Ignitor.cpp
@@ -1,5 +1,5 @@
 /* 
- * Ignitor.cpp - pump control
+ * Ignitor.cpp - ignitor control
  */
 
 #include "application.h"
@@ -12,7 +12,6 @@ Ignitor::Ignitor(int pin) {
 }
 
 void Ignitor::fire() {
-    // turn ignitor on for 5 sec and record time when that occurs
     digitalWrite(_pin, HIGH);
     delay(_ignitor_on_time);
     digitalWrite(_pin, LOW);