+++ software/ADPL_electron/Valve.h
@@ -11,6 +11,7 @@ class Valve {
         bool gasOn;
     private:
         int _pin;
+        unsigned long _delay_before_ignition = 100;  // add some delay before igniting
 };
 
 #endif