+++ software/ADPL_electron/error_codes.h
@@ -1,0 +1,8 @@
+// Define error codes for log.error()
+#define SD_INIT_FAIL 1
+#define CELL_PUB_FAIL 2
+#define SD_FILE_REMOVE_FAIL 3
+#define SD_DATA_PUSH_FAIL 4
+#define SD_PUB_FAIL 5
+#define TOTAL_PUB_FAIL 6
+#define IGNITOR_FAIL 7