+++ software/ADPL_electron/pin_mapping.h
@@ -1,18 +1,22 @@
 // Define GPIO pins (Particle Electron) by function
 #define HXCI A0 // Heat Exchanger Cold Inlet
-#define HXCO A1 // Heat Exchanger Cold Outlet
+#define HXCO A1 // Heat Exchanger Cold Output
 #define HTR A2  // Heater
 #define HXHI A3 // Heat Exchanger Hot Inlet
-#define HXHO A4 // Heat Exchanger Hot Outlet
-#define VALVE D4
-#define IGNITOR D2
+#define HXHO A4 // Heat Exchanger Hot Outlet / Changed to A5 from A4 pin burnt
+#define EXH A5 // Heater Exhaust
 #define PUMP D0
-//#define LEVEL B5
-//#define BUCKET C0 // Duke
-#define BUCKET C4 // Kenya
-#define STEP D3
+#define VALVE D1
+#define IGNITOR D2
+#define SLEEP D3
+#define STEP D6
 #define DIR D5
-#define SLEEP D6
 #define UP B5
-#define DOWN B4
+#define DOWN C3
 #define RESET C2
+#define BUCKET C4
+#define SD_CD_PIN B4
+#define SD_CS_PIN B3
+#define SD_DI_PIN B2
+#define SD_DO_PIN B1
+#define SD_CLK_PIN B0