+++ software/ADPL_electron/CHANGELOG.md
@@ -5,7 +5,17 @@ pre-electron-v2.1.0
 -------------------
 * See the git commit history / tagged comments for notes.
 
-v2.2.0
-------
+electron_v2.2.0
+---------------
 * Drop the 'electron' prefix since everything is electron-based now.
 * Remove level-sensing.
+
+electron_v2.4.0a
+----------------
+* Daniel's pinch valve functionality
+
+v2.5.0RC1
+----------
+* Change pinch valve functionality to binary mode for unclogs
+* SD card functionality using SdFat
+* Exhaust thermocouple for ignitor control